Removing reference to the old project.
I was asked where I got this information from. It is NOT illegally
copied from the SDK, all the info here is obtained from example
programs and/or documentation readily available on the internet.

An example : Use the URL
http://msdn.microsoft.com/library/default.asp?url=/library/en-us/dnanchor/html/windowsce.asp
and do a search on SHCreateMenuBar, then select the third result
(the one from the PocketPC Developer's Reference). You'll see a
function declaration there. Then follow the link SHMENUBARINFO
and you'll see more detail.