This isn't exactly clean yet, but I want to save a somewhat working version of this before messing it up further :-)
This adds an ARM version of the mcount function so it is possible to
do profiling. It also ports from the use of unix's open/write/close
system calls to use of stdio (which CE's coredll also has), and
replaces calls to perror and use of errno by ugly constructions.

In this state, it can be made to work under arm-wince-mingw32ce.