Submitted by Nuno Lucas.
The GetCharWidth32 hasn't two variations on WinCE as it has on Windows
(even if they are the same).

> This confuses me. Should we define MessageBoxW in the lib with the W,
> but GetCharWidth32 without the W ? What do other Windows architectures
> implement ?

I believe this was actually an error by Microsoft on the Windows API,
which they maintained for the sake of compatibility. But with Windows
CE they didn't need to maintain that error and just had one
GetCharWidth32() export.